requirements.txt
discord.py~=1.7.1
asyncpraw~=7.2.0